Tests for settings and more

+++ server/services/shippingpackage.ts
@@ -2,19 +2,34 @@ import { Strapi } from '@strapi/strapi';
 
 export default ({ strapi }: { strapi: Strapi }) => ({
 
-  async find(query) {
-    return await strapi.entityService.findMany("plugin::omcommerce.shippingpackage", query);
+  async find(query: any) {
+    if (strapi.entityService) {
+      // @ts-ignore
+      return await strapi.entityService.find("plugin::omcommerce.shippingpackage", query);
+    } else {
+      throw new Error('strapi.entityService is not defined');
+    }
+  },
+  async create(data: any) {
+    if (strapi.entityService) {
+      return await strapi.entityService.create("plugin::omcommerce.shippingpackage", data);
+    } else {
+      throw new Error('strapi.entityService is not defined');
+    }
   },
 
-
-  async create(data) {
-    return await strapi.entityService.create("plugin::omcommerce.shippingpackage", data);
+  async update(id: any, data: any) {
+    if (strapi.entityService) {
+      return await strapi.entityService.update("plugin::omcommerce.shippingpackage", id, data);
+    } else {
+      throw new Error('strapi.entityService is not defined');
+    }
   },
-
-  async update(id, data) {
-    return await strapi.entityService.update("plugin::omcommerce.shippingpackage", id, data);
-  },
-  async delete(id) {
-    return await strapi.entityService.delete("plugin::omcommerce.shippingpackage", id);
+  async delete(id: any) {
+    if (strapi.entityService) {
+      return await strapi.entityService.delete("plugin::omcommerce.shippingpackage", id);
+    } else {
+      throw new Error('strapi.entityService is not defined');
+    }
   },
 });
